Restructured how appointments are booked
Appointments are now assigned a 'leader', which is the first person to book that appointment. This confirmation code will be given to the leader, and anyone who wants to book the same appointment will need that ocnfirmation code to do so. If the leader cancels an appointment, the person who booked the appointment the earliest becomes the new leader, and a new confirmation code is set. This change required creation of a new route called 'book_existing_appointment' to book an appointment that already has a leader. It also required changing the database so that only the leader can set an appointment subject and location.