Changed how available appointments are requested
Available appointments are now requested by month instead of by day to minimize fetch requests to the backend. This allows the Calendar to store all available appointments for the month. That change means that AppointmentList no longer needs to be a child component that requests appointment info, and it can simply be a component that accepts props from Calendar.